implement time zone differently
Combine Date, Time and DateAndTime modules, and move TimeZone to the new
module, so that TimeZone can work without exposing functions we do not
want exposed (because of potential errors involved).

TimeZone is now completely independent of utc. It is defined as periods,
with each period defining a mapping between moment and date/time. This
is something very new to date/time libraries, but when I started
thinking this way, everything got way simpler.